pyproject.toml
[project]
name = "recursivegarden"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.11",
    "ipdb>=0.13.13",
    "jinja2>=3.1.6",
    "llm>=0.23",
    "llm-anthropic>=0.15.1",
    "pillow>=11.1.0",
    "pydantic-xml>=2.14.3",
    "python-frontmatter>=1.1.0",
    "python-multipart>=0.0.20",
    "pyyaml>=6.0.2",
    "ruff>=0.9.6",
    "uvicorn>=0.34.0",
]
